% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_impulses.R
\name{compute_impulses}
\alias{compute_impulses}
\title{Compute impulse responses}
\usage{
compute_impulses(bh, swish, nn)
}
\arguments{
\item{bh}{A draw of the parameter matrix.}

\item{swish}{A draw of the covariance matrix.}

\item{nn}{A vector containing the amount of variables in the system, number
of lags and prediction horizon.}
}
\value{
A matrix with impulse responses
}
\description{
Take the draws of parameter matrix and covariance matrix and compute impulse
responses for a predefined number of horizon.
}
\references{
Code inspired by compute_impulses function of package VARsignR
by Christian Danne (2015)
}
