% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebr_accept.R
\name{ebr_accept}
\alias{ebr_accept}
\title{Check whether elasticity bounds are satisfied}
\usage{
ebr_accept(ebr = NULL, impulses = NULL, Q = NULL, allshocknames, data = data)
}
\arguments{
\item{ebr}{Object of ebr_setup.}

\item{impulses}{Computed impulses based on parameter draws.}

\item{Q}{A rotation matrix.}

\item{allshocknames}{Vector of all shocknames which are identified in the
the system.}

\item{data}{A data frame or tibble containg the data}
}
\value{
A Boolean whether elasticity bounds are fulfilled
}
\description{
ebr_accept takes values for IRFs which are based on current draws, computes
elasticity and assesses whether the elasticity bound restriction is
satisified. Note that elasticities are computed here as fractions. This
approach is likely only appropriate for a limited amount of application.
}
\references{
Diaz Ramirez (2018)
}
