% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restriction_check.R
\name{restriction_check}
\alias{restriction_check}
\title{Check restrictions for NarrSign model}
\usage{
restriction_check(
  tradsign_setup_obj = NULL,
  ebr_setup_obj = NULL,
  narrsign_obj = NULL
)
}
\arguments{
\item{tradsign_setup_obj}{output of tradsign_setup.}

\item{ebr_setup_obj}{output of ebr_setup.}

\item{narrsign_obj}{output of narrsign_setup.}
}
\value{
A print statement summarising imposed restrictions
}
\description{
restriction_check prints out a verbal summary of the imposed restrictions.
}
\examples{
\dontrun{
restriction_check(tradsign_setup,
ebr_setup,
narrsign_setup)
}
}
