% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uhlig}
\alias{uhlig}
\title{Uhlig replication data}
\format{
A tibble with 515 Rows and 7 variables:

\describe{
    \item{dates}{Monthly dates in yyyy-mm-dd format}
    \item{gdpc1}{GDP growth}
    \item{gpddef}{Implicit price deflator}
    \item{cprindex}{Commodity price index}
    \item{totresns}{Total reserves}
    \item{bogonbr}{Unborrowed reserves}
    \item{fedfunds}{Federal Funds rate}
}
}
\source{
replication data Diaz Ramirez et al. (2018)
}
\usage{
uhlig
}
\description{
A dataset containing one column of dates and 6 variables.
}
\references{
Uhlig 2005
}
\keyword{datasets}
