% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_shock_plot.R
\name{two_shock_plot}
\alias{two_shock_plot}
\title{Plot structural shocks for all identified shocks}
\usage{
two_shock_plot(narrSign_model = NULL, whichShock = NULL, prob = 0.5)
}
\arguments{
\item{narrSign_model}{A model object of class narrsign}

\item{whichShock}{A string with the name of shock to be plotted or a numeric
with the position of the shock within the system}

\item{prob}{Argument for quantile function}
}
\value{
A ggplot object
}
\description{
Plot structural shocks for all identified shocks
}
